Sanitize incoming EXIF before Postgres insert.
REST creates were storing IPTC ApplicationRecordVersion with a null byte, which JSON/JSONB rejects. Run the existing sanitizer on gallery-images and media beforeChange, not only in the EXIF job.
